requirements.txt
fastapi
uvicorn
streamlit
asyncpg
psycopg2-binary
redis
elasticsearch
pymilvus
sentence-transformers
openai
httpx
tiktoken
pandas<2.3.0
numpy<2.3.0
PyPDF2
python-docx
mammoth
pdf2image
Pillow
python-dotenv
python-multipart
pydantic
plotly==5.18.0
duckduckgo-search
h11>=0.16.0
flashrank